+++ packages/client/src/Config.ts
@@ -12,7 +12,7 @@ import CONFIG_SCHEMA from './config.schema.json'
 import { SmartContractRecord } from 'streamr-client-protocol'
 import { LogLevel } from '@streamr/utils'
 
-import { PRODUCTION_STUN_URLS, NetworkNodeOptions } from 'streamr-network'
+import { NetworkNodeOptions, STREAMR_ICE_SERVERS } from 'streamr-network'
 import type { ConnectionInfo } from '@ethersproject/web'
 import { generateClientId } from './utils/utils'
 
@@ -205,6 +205,28 @@ export const createStrictConfig = (inputOptions: StreamrClientConfig = {}): Stri
     const opts = cloneDeep(inputOptions)
     const defaults = cloneDeep(STREAM_CLIENT_DEFAULTS)
 
+    const getMetricsConfig = () => {
+        if (opts.metrics === true) {
+            return defaults.metrics
+        } else if (opts.metrics === false) {
+            return {
+                ...defaults.metrics,
+                periods: []
+            }
+        } else if (opts.metrics !== undefined) {
+            return {
+                ...defaults.metrics,
+                ...opts.metrics
+            }
+        } else {
+            const isEthereumAuth = (opts.auth?.ethereum !== undefined)
+            return {
+                ...defaults.metrics,
+                periods: isEthereumAuth ? [] : defaults.metrics.periods
+            }
+        }
+    }
+
     const options: StrictStreamrClientConfig = {
         id: generateClientId(),
         ...defaults,
@@ -214,17 +236,7 @@ export const createStrictConfig = (inputOptions: StreamrClientConfig = {}): Stri
             trackers: opts.network?.trackers ?? defaults.network.trackers,
         },
         decryption: merge(defaults.decryption || {}, opts.decryption),
-        metrics: (opts.metrics === true)
-            ? defaults.metrics
-            : (opts.metrics === false)
-                ? {
-                    ...defaults.metrics,
-                    periods: []
-                }
-                : {
-                    ...defaults.metrics,
-                    ...opts.metrics
-                },
+        metrics: getMetricsConfig(),
         cache: {
             ...defaults.cache,
             ...opts.cache,
@@ -239,8 +251,8 @@ export const createStrictConfig = (inputOptions: StreamrClientConfig = {}): Stri
         }
     }
 
-    if (options.network.stunUrls === undefined) {
-        options.network.stunUrls = PRODUCTION_STUN_URLS
+    if (options.network.iceServers === undefined) {
+        options.network.iceServers = STREAMR_ICE_SERVERS
     }
 
     return options