+++ packages/client/src/Stream.ts
@@ -7,7 +7,6 @@ import { StreamRegistry } from './registry/StreamRegistry'
 import { StreamRegistryCached } from './registry/StreamRegistryCached'
 import {
     StreamID,
-    StreamMessage,
     StreamPartID,
     toStreamPartID
 } from 'streamr-client-protocol'
@@ -17,13 +16,17 @@ import { PermissionAssignment, PublicPermissionQuery, UserPermissionQuery } from
 import { Subscriber } from './subscribe/Subscriber'
 import { formStorageNodeAssignmentStreamId } from './utils/utils'
 import { waitForAssignmentsToPropagate } from './utils/waitForAssignmentsToPropagate'
-import { MessageMetadata } from '../src/publish/Publisher'
+import { PublishMetadata } from '../src/publish/Publisher'
 import { StreamStorageRegistry } from './registry/StreamStorageRegistry'
 import { toEthereumAddress, withTimeout } from '@streamr/utils'
 import { StreamMetadata } from './StreamMessageValidator'
 import { StreamrClientEventEmitter } from './events'
 import { collect } from './utils/iterators'
 import { DEFAULT_PARTITION } from './StreamIDBuilder'
+import { Subscription } from './subscribe/Subscription'
+import { LoggerFactory } from './utils/LoggerFactory'
+import { Message } from './Message'
+import { convertStreamMessageToMessage } from './Message'
 
 export interface StreamProperties {
     id: string
@@ -86,6 +89,7 @@ class StreamrStream implements StreamMetadata {
     private readonly _streamRegistry: StreamRegistry
     private readonly _streamRegistryCached: StreamRegistryCached
     private readonly _streamStorageRegistry: StreamStorageRegistry
+    private readonly _loggerFactory: LoggerFactory
     private readonly _eventEmitter: StreamrClientEventEmitter
     private readonly _timeoutsConfig: TimeoutsConfig
 
@@ -98,6 +102,7 @@ class StreamrStream implements StreamMetadata {
         streamRegistryCached: StreamRegistryCached,
         streamRegistry: StreamRegistry,
         streamStorageRegistry: StreamStorageRegistry,
+        loggerFactory: LoggerFactory,
         eventEmitter: StreamrClientEventEmitter,
         timeoutsConfig: TimeoutsConfig
     ) {
@@ -110,6 +115,7 @@ class StreamrStream implements StreamMetadata {
         this._streamRegistryCached = streamRegistryCached
         this._streamRegistry = streamRegistry
         this._streamStorageRegistry = streamStorageRegistry
+        this._loggerFactory = loggerFactory
         this._eventEmitter = eventEmitter
         this._timeoutsConfig = timeoutsConfig
     }
@@ -166,9 +172,9 @@ class StreamrStream implements StreamMetadata {
 
         if (!receivedMsgs.length) { return }
 
-        const lastMessage = receivedMsgs[0].getParsedContent()
+        const lastMessage = receivedMsgs[0].content
 
-        const fields = Object.entries(lastMessage).map(([name, value]) => {
+        const fields = Object.entries(lastMessage as any).map(([name, value]) => {
             const type = getFieldType(value)
             return !!type && {
                 name,
@@ -192,7 +198,8 @@ class StreamrStream implements StreamMetadata {
         const normalizedNodeAddress = toEthereumAddress(nodeAddress)
         try {
             const streamPartId = toStreamPartID(formStorageNodeAssignmentStreamId(normalizedNodeAddress), DEFAULT_PARTITION)
-            assignmentSubscription = await this._subscriber.add(streamPartId)
+            assignmentSubscription = new Subscription<any>(streamPartId, this._loggerFactory)
+            await this._subscriber.add(assignmentSubscription)
             const propagationPromise = waitForAssignmentsToPropagate(assignmentSubscription, this)
             await this._streamStorageRegistry.addStreamToStorageNode(this.id, normalizedNodeAddress)
             await withTimeout(
@@ -225,11 +232,10 @@ class StreamrStream implements StreamMetadata {
     /**
      * @category Important
      */
-    async publish<T>(content: T, metadata?: MessageMetadata): Promise<StreamMessage<T>> {
-        const result = this._publisher.publish(this.id, content, metadata)
+    async publish<T>(content: T, metadata?: PublishMetadata): Promise<Message> {
+        const result = await this._publisher.publish(this.id, content, metadata)
         this._eventEmitter.emit('publish', undefined)
-        return result
-
+        return convertStreamMessageToMessage(result)
     }
 
     /** @internal */