+++ packages/client/src/index-exports.ts
@@ -3,9 +3,10 @@
  */
 export * from './StreamrClient'
 export * from './Stream'
+export { Message, MessageMetadata } from './Message'
 export { DecryptError } from './encryption/EncryptionUtil'
 export { StreamrClientEvents } from './events'
-export { MessageMetadata } from './publish/Publisher'
+export { PublishMetadata } from './publish/Publisher'
 export { Subscription, SubscriptionEvents, } from './subscribe/Subscription'
 export type { MessageStream, MessageListener } from './subscribe/MessageStream'
 export { ResendOptions, ResendLastOptions, ResendFromOptions, ResendRangeOptions, ResendRef } from './subscribe/Resends'