+++ packages/client/src/subscribe/ResendSubscription.ts
@@ -4,7 +4,6 @@ import { StreamMessage, StreamPartID, StreamPartIDUtils } from 'streamr-client-p
 import { ConfigInjectionToken } from '../Config'
 import { OrderMessages } from './OrderMessages'
 import { ResendOptions, Resends } from './Resends'
-import { DestroySignal } from '../DestroySignal'
 import { LoggerFactory } from '../utils/LoggerFactory'
 import { SubscribeConfig } from './../Config'
 import { MessageStream } from './MessageStream'
@@ -17,25 +16,20 @@ export class ResendSubscription<T> extends Subscription<T> {
         streamPartId: StreamPartID,
         private resendOptions: ResendOptions,
         private resends: Resends,
-        destroySignal: DestroySignal,
         loggerFactory: LoggerFactory,
         @inject(ConfigInjectionToken.Subscribe) subscibreConfig: SubscribeConfig
     ) {
         super(streamPartId, loggerFactory)
-        this.resendThenRealtime = this.resendThenRealtime.bind(this)
         this.orderMessages = new OrderMessages<T>(
             subscibreConfig,
             resends,
             streamPartId,
             loggerFactory
         )
-        this.pipe(this.resendThenRealtime)
+        this.pipe(this.resendThenRealtime.bind(this))
         this.pipe(this.orderMessages.transform())
         this.onBeforeFinally.listen(async () => {
             this.orderMessages.stop()
-        })
-        destroySignal.onDestroy.listen(() => {
-            this.eventEmitter.removeAllListeners()
         })
     }
 
@@ -52,7 +46,7 @@ export class ResendSubscription<T> extends Subscription<T> {
 
     private async* resendThenRealtime(src: AsyncGenerator<StreamMessage<T>>): AsyncGenerator<StreamMessage<T>, void, any> {
         try {
-            yield* await this.getResent()
+            yield* (await this.getResent()).getStreamMessages()
         } catch (err) {
             if (err.code === 'NO_STORAGE_NODES') {
                 const streamId = StreamPartIDUtils.getStreamID(this.streamPartId)