+++ packages/network/src/connection/webrtc/NodeWebRtcConnection.ts
@@ -116,7 +116,12 @@ export class NodeWebRtcConnection extends WebRtcConnection {
 
     protected doConnect(): void {
         this.connection = new nodeDataChannel.PeerConnection(this.selfId, {
-            iceServers: [...this.stunUrls],
+            iceServers: this.iceServers.map(({ url, port, username, password }) => ({
+                hostname: url,
+                port,
+                username,
+                password
+            })),
             maxMessageSize: this.maxMessageSize
         })